Import FileHandler class to Records class to use static methods
